cleanup of code moves
